feat: DetailActivity 추가
