+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,11 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
+
+        <activity
+            android:name=".presentation.DetailActivity"
+            android:exported="true">
+        </activity>
     </application>
 
 </manifest>
