+++ app/src/main/java/com/survivalcoding/noteapp/presentation/DetailActivity.kt
@@ -1,0 +1,14 @@
+package com.survivalcoding.noteapp.presentation
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.survivalcoding.noteapp.databinding.ActivityDetailBinding
+
+class DetailActivity : AppCompatActivity() {
+    private val binding: ActivityDetailBinding by lazy { ActivityDetailBinding.inflate(layoutInflater) }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+    }
+}
