+++ app/src/main/java/com/survivalcoding/noteapp/presentation/fragment/AddFragment.kt
@@ -6,16 +6,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import com.survivalcoding.noteapp.databinding.FragmentEditBinding
+import com.survivalcoding.noteapp.databinding.FragmentAddBinding
 import com.survivalcoding.noteapp.presentation.MainViewModel
 
-class EditFragment : Fragment() {
-    private var _binding: FragmentEditBinding? = null
+class AddFragment : Fragment() {
+    private var _binding: FragmentAddBinding? = null
     private val binding get() = _binding!!
     private val viewModel: MainViewModel by activityViewModels()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        _binding = FragmentEditBinding.inflate(inflater, container, false)
+        _binding = FragmentAddBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         return root
