--- app/src/main/java/com/survivalcoding/noteapp/presentation/fragment/MainFragment.kt
@@ -1,47 +1,0 @@
-package com.survivalcoding.noteapp.presentation.fragment
-
-import android.os.Bundle
-import android.view.*
-import androidx.core.view.MenuProvider
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.Lifecycle
-import com.survivalcoding.noteapp.R
-import com.survivalcoding.noteapp.databinding.FragmentMainBinding
-import com.survivalcoding.noteapp.presentation.MainViewModel
-
-class MainFragment : Fragment() {
-    private var _binding: FragmentMainBinding? = null
-    private val binding get() = _binding!!
-    private val viewModel: MainViewModel by activityViewModels()
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        _binding = FragmentMainBinding.inflate(inflater, container, false)
-        val root: View = binding.root
-
-        setupMenu()
-
-        return root
-    }
-
-    private fun setupMenu() {
-        binding.toolbar.addMenuProvider(object : MenuProvider {
-            override fun onPrepareMenu(menu: Menu) {
-
-            }
-
-            override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
-                menuInflater.inflate(R.menu.menu_main, menu)
-            }
-
-            override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
-                when (menuItem.itemId) {
-                    R.id.sort -> {
-                        // TODO
-                    }
-                }
-                return true
-            }
-        }, viewLifecycleOwner, Lifecycle.State.STARTED)
-    }
-}
