+++ app/src/main/res/layout/activity_detail.xml
@@ -4,15 +4,20 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@android:color/black">
+    android:background="@color/black"
+    tools:context=".presentation.DetailActivity">
 
-    <TextView
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/save_fab"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="hello world!"
-        android:textColor="@color/white"
+        android:layout_margin="@dimen/margin_default"
+        android:backgroundTint="@android:color/white"
+        android:contentDescription="@string/description_fab"
+        android:elevation="0dp"
+        android:src="@drawable/ic_save"
+        app:elevation="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
